Finished paint-app/django integration
The paint functionality works and the server can successfully retrieve
the data URL from the user's input via a django-created HTML form.
Included in this commit is an export of the newest version of the model.
What remains is to connect the two. In particular, ideally the data URL
would be read from memory.