Saving basically no progress
For future research: There is an error where one of my packages depends
on beniget 0.4.1 which itself depends on gast=~0.5.0 (which means
gast>=0.5.0, gast==0.5.*). This gast package is declared as a dependency
but as gast==0.4.0, which is the maximum version tensorflow supports.

The nixpkgs version of gast (pkgs.python310Packages.gast) is at version
0.5.3 or so, which satisfies.

Trying to override gast to >=0.5.0 by setting it as
gast=pkgs.python310Packages.gast doesn't work for the same error
message. This suggests to me that it has something to do with the
"internal" build process of beniget- however...

Trying to add pkgs.python310Packages.gast to buildInputs or
nativeBuildInputs of beniget fails with an error saying that I am trying to install
two versions of a package simultaneously. This means then, that either
there is something irreconcilable with gast=0.4.0 in the TOML (it is taking the
version declared in the TOML in) or something else is going wrong.

Future steps: could try to output the buildInputs or
nativeBuildInputs or whatever of the beniget package. My current way of
"overriding" the inputs just inserts more elements into the list- if a
version of gast is already in there (to see that I think I would have to
print it out) then that could be the source of my error.