Forcing poetry to create-envs (not sure if useful)
The reasoning for this is that sometimes when changing the python
version to one which is out of bounds of the currently activated
.venv (and maybe under some other equally special circumstances), poetry
install will refuse to build a new venv.

The hope is that forcing it to do so will make it install venvs "no
matter what", thus overriding the current venv if need be.