Dev fixes
Poetry thru poetry2nix is stuck at an old version, and that contributed
to breakages with another project. Since this flake is general, I'm
pinning it further upstream.

Pyright was causing weird errors. In hindsight, this was probably
because of the issue with file watchers I committed against today.

Either way, it now exists in the dev environment two ways, for whatever reason.