Black magic (revert if tensorflow explodes)
Ok, so, basically the issue with tensorflow is that it looks for a
version of the standard c library globally.

- on nixos, because of how it works, this library is not installed in
the usual place, so it complains about missing a .so file

- overriding LD_LIBRARY_PATH violates assumptions for normal nix
packages
for example, pkgsOld.python[...] uses it. so if you override it to
something that leaves out one of its library paths, things blow up.

- additionally, there is an outstanding issue where overriding
LD_LIBRARY can sometimes blow up nix itself

- ${pkgs}.cc.cc is the path to the c compiler (and libs) provided in the
stdenv used to build all derivations in the revision of ${pkgs}

- makeLibraryPath converts this path into an actual filesystem path to
include in LD_LIBRARY (it's basically in packageloc/lib)

- so, like NobbZ said a few weeks ago, the problem is truly operating
system specific:

+ on normal linux distros, it's probably best to leave out LD_LIBRARY_PATH
in general, since the c compiler is already stored in a normal place

+ on NixOS, you should use stdenv.cc.cc. the version of nixpkgs you
should use probably depends